moved models to models.py
