+++ forms.py
@@ -1,5 +1,5 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, SubmitField, PasswordField, SelectField
+from wtforms import StringField, SubmitField, PasswordField, SelectField, IntegerField
 from wtforms.validators import DataRequired, URL
 
 # RegisterForm to register new users
@@ -18,3 +18,16 @@ class LoginForm(FlaskForm):
     email = StringField('Email Address', validators=[DataRequired()])
     password = PasswordField('Password', validators=[DataRequired()])
     submit = SubmitField('Login')
+
+# Used for adding customer ranking, rating and review; Used within Edit route
+class RateMovieForm(FlaskForm):
+    # DataRequired() ensures user completes that section
+    rating = StringField('Your rating out of 10 e.g. 7.5', validators=[DataRequired()])
+    review = StringField('Your Review', validators=[DataRequired()])
+    ranking = IntegerField('Your new ranking', validators=[DataRequired()])
+    submit = SubmitField("Update")
+
+# WTForm for Movie API search. Gets called @add_movie
+class AddMovie(FlaskForm):
+    title = StringField('Movie Title', validators=[DataRequired()])
+    sumit = SubmitField("Add Movie")
