+++ templates/register.html
@@ -1,3 +1,11 @@
 {% extends 'base.html' %}
+{% from 'bootstrap5/form.html' import render_form %}
 
 {% block title %}Register{% endblock %}
+
+{% block content %}
+<div class="content">
+  <h1 class="heading">Register User</h1>
+    {{ render_form(form) }}
+  </div>
+{% endblock %}
